requirements.txt
streamlit
rembg
pillow
watchdog
numpy
onnxruntime